requirements.txt
bleach==1.5.0
click==6.7
Flask==0.12.2
Flask-PyMongo==0.5.1
google-api-python-client==1.6.3
html5lib==0.9999999
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.6
Keras==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
numpy==1.13.1
oauth2client==4.1.2
pandas==0.20.3
protobuf==3.4.0
pyasn1==0.3.5
pyasn1-modules==0.1.4
pymongo==3.5.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
rsa==3.4.2
scipy==0.19.1
six==1.10.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.6
uritemplate==3.0.0
Werkzeug==0.12.2
